Lots of fixes and improvements.
1: Found some dialogs available on videos and sites. Tried my best to copy them.
2: Removed a function "getTradeItems" that wasn't being used.
3: Relocated and improved the trade function.
4: Fixed Reviewdog's warnings of lines bigger than 120 characters.

+++ data/npc/scripts/grizzly_adams.lua
@@ -17,11 +17,9 @@ end
 
 local choose = {}
 local cancel = {}
-local available = {}
 
 local grizzlyAdamsConfig = {
 	ranks = {
-		--NOTE: The variable 'name' is not necessary to be declared. I let it so people who wants to change the script will now wich item is each one.
 		huntsMan_rank = {
 			-- SELL OFFERS
 			{id=11214, buy=0, sell=50, name="antlers"},
@@ -105,20 +103,13 @@ local function greetCallback(cid)
 	local player = Player(cid)
 	if player:getStorageValue(JOIN_STOR) == -1 then
 		npcHandler:setMessage(MESSAGE_GREET,
-					"Welcome |PLAYERNAME|. Would you like to join the 'Paw and Fur - Hunting Elite'?")
+					"Hi there, do you want to to join the 'Paw and Fur - Hunting Elite'?")
 	else
-		npcHandler:setMessage(MESSAGE_GREET, "Welcome back old chap. What brings you here this time?")
+		npcHandler:setMessage(MESSAGE_GREET,
+							"Welcome to the 'Paw and Fur - Hunting Elite' |PLAYERNAME|. "..
+							"Feel free to do {tasks} for us.")
 	end
 	return true
-end
-
-local function getTradeItems(t)
-	local list, obj = {}
-	for i = 1, #t do
-		obj = t[i]
-		list[obj.id] = {id = obj.id, buy = obj.buy, sell = obj.sell, name = ItemType(obj.id):getName():lower()}
-	end
-	return list
 end
 
 local function joinTables(old, new)
@@ -159,6 +150,22 @@ local function onSell(cid, item, subType, amount, ignoreCap, inBackpacks)
 	return true
 end
 
+local function startTrade(cid, player)
+	if player:getPawAndFurRank() >= 2 then
+		local tradeItems = grizzlyAdamsConfig.ranks.huntsMan_rank
+		if player:getPawAndFurRank() >= 4 then
+			tradeItems = joinTables(tradeItems, grizzlyAdamsConfig.ranks.bigGameHunter_rank)
+		end
+		if player:getPawAndFurRank() >= 6 then
+			tradeItems = joinTables(tradeItems, grizzlyAdamsConfig.ranks.trophyHunter_rank)
+		end
+		openShopWindow(cid, tradeItems, onBuy, onSell)
+		return npcHandler:say("It's my offer.", cid)
+	else
+		return npcHandler:say("You don't have any rank.", cid)
+	end
+end
+
 local function creatureSayCallback(cid, type, msg)
 	if not npcHandler:isFocused(cid) then
 		return false
@@ -166,21 +173,9 @@ local function creatureSayCallback(cid, type, msg)
 
 	local player = Player(cid)
 	msg = msg:gsub("(%l)(%w*)", function(a,b) return string.upper(a)..b end)
+
 	if msgcontains("trade", msg) then
-		if player:getPawAndFurRank() >= 2 then
-			local tradeItems = {}
-			tradeItems = grizzlyAdamsConfig.ranks.huntsMan_rank
-			if player:getPawAndFurRank() >= 4 then
-				tradeItems = joinTables(tradeItems, grizzlyAdamsConfig.ranks.bigGameHunter_rank)
-			end
-			if player:getPawAndFurRank() >= 6 then
-				tradeItems = joinTables(tradeItems, grizzlyAdamsConfig.ranks.trophyHunter_rank)
-			end
-			openShopWindow(cid, tradeItems, onBuy, onSell)
-			return npcHandler:say("It's my offer.", cid)
-		else
-			return npcHandler:say("You don't have any rank.", cid)
-		end
+		startTrade(cid, player)
 	elseif (msgcontains("join", msg) or msgcontains("yes", msg))
 			and npcHandler.topic[cid] == 0
 			and player:getStorageValue(JOIN_STOR) ~= 1 then
@@ -189,10 +184,12 @@ local function creatureSayCallback(cid, type, msg)
 					"A warm welcome to our newest member: |PLAYERNAME|! " ..
 					"Ask me for a {task} if you want to go on a hunt.", cid)
 	elseif isInArray({"tasks", "task", "mission"}, msg:lower()) then
-		local can = player:getTasks()
 		if player:getStorageValue(JOIN_STOR) == -1 then
 			return npcHandler:say("You'll have to {join}, to get any {tasks}.",cid)
 		end
+
+		local can = player:getTasks()
+
 		if #can > 0 then
 			local text = ""
 			local sep = ", "
@@ -209,46 +206,46 @@ local function creatureSayCallback(cid, type, msg)
 				end
 				text = text .. "{" .. (tasks[id].name or tasks[id].raceName) .. "}" .. sep
 			end
-			npcHandler:say("The current task" .. (#can > 1 and "s" or "") .. " that you can choose " .. (#can > 1 and "are" or "is") .. " " .. text, cid)
+			npcHandler:say("The current task" ..
+						(#can > 1 and "s" or "") .. " that you can choose " ..
+						(#can > 1 and "are" or "is") .. " " .. text, cid)
 			npcHandler.topic[cid] = 0
 		else
 			npcHandler:say("I don't have any task for you right now.", cid)
 		end
 	elseif msg ~= "" and player:canStartTask(msg) then
 		if #player:getStartedTasks() >= tasksByPlayer then
-			npcHandler:say("Sorry, but you already started " .. tasksByPlayer .. " tasks. You can check their {status}, {cancel} or {report} a task.", cid)
+			npcHandler:say("Sorry, but you already started " .. tasksByPlayer .. " tasks."..
+						" ".."You can check their {status}, {cancel} or {report} a task.", cid)
 			return true
 		end
 		local task = getTaskByName(msg)
 		if task and player:getStorageValue(QUESTSTORAGE_BASE + task) > 0 then
 			return false
 		end
-		npcHandler:say("In this task you must defeat " .. tasks[task].killsRequired .. " " .. tasks[task].raceName .. ". Are you sure that you want to start this task?", cid)
+		npcHandler:say("In this task you must defeat " .. tasks[task].killsRequired .. " " .. tasks[task].raceName .. "."..
+					" ".."Are you sure that you want to start this task?", cid)
 		choose[cid] = task
 		npcHandler.topic[cid] = 1
 	elseif msg:lower() == "yes" and npcHandler.topic[cid] == 1 then
 		player:setStorageValue(QUESTSTORAGE_BASE + choose[cid], 1)
 		player:setStorageValue(KILLSSTORAGE_BASE + choose[cid], 0)
-		npcHandler:say("Excellent! You can check the {status} of your task saying {report} to me. Also you can {cancel} tasks to.", cid)
+		npcHandler:say("Excellent! You can check the {status} of your task saying {report} to me."..
+					" ".."Also you can {cancel} tasks to.", cid)
 		choose[cid] = nil
 		npcHandler.topic[cid] = 0
 	elseif msgcontains("status", msg) then
 		local started = player:getStartedTasks()
 		if started and #started > 0 then
 			local text = ""
-			local sep = ", "
 			table.sort(started, (function(a, b) return (a < b) end))
 			local t = 0
 			local id
 			for i = 1, #started do
 				id = started[i]
 				t = t + 1
-				if t == #started - 1 then
-					sep = " and "
-				elseif t == #started then
-					sep = "."
-				end
-				text = text .. "Task name: " .. tasks[id].raceName .. ". Current kills: " .. player:getStorageValue(KILLSSTORAGE_BASE + id) .. ".\n"
+				text = text .. "Task name: " .. tasks[id].raceName .. ". "..
+					"Current kills: " .. player:getStorageValue(KILLSSTORAGE_BASE + id) .. ".\n"
 			end
 			npcHandler:say({"The status of your current tasks is:\n" .. text}, cid)
 		else
@@ -316,10 +313,13 @@ local function creatureSayCallback(cid, type, msg)
 						end
 						text = text .. "{" .. (tasks[id].name or tasks[id].raceName) .. "}" .. sep
 					end
-					npcHandler:say("The current task" .. (#started > 1 and "s" or "") .. " that you started " .. (#started > 1 and "are" or "is") .. " " .. text, cid)
+					npcHandler:say("The current task" .. (#started > 1 and "s" or "") ..
+								" that you started " .. (#started > 1 and "are" or "is") .. " " .. text, cid)
 				end
 			else
-				npcHandler:say("Awesome! you finished " .. (finished > 1 and "various" or "a") .. " task" .. (finished > 1 and "s" or "") .. ". Talk to me again if you want to start a {task}.", cid)
+				npcHandler:say("Awesome! "..
+							"You finished "..(finished > 1 and "various" or "a").." task"..(finished > 1 and "s" or "").."."..
+							"Talk to me again if you want to start a {task}.", cid)
 			end
 		else
 			npcHandler:say("You haven't started any task yet.", cid)
@@ -343,13 +343,13 @@ local function creatureSayCallback(cid, type, msg)
 				text = text .. "{" .. (tasks[id].name or tasks[id].raceName) .. "}" .. sep
 			end
 
-			npcHandler:say("The current task" .. (#started > 1 and "s" or "") .. " that you started " .. (#started > 1 and "are" or "is") .. " " .. text, cid)
+			npcHandler:say("The current task" .. (#started > 1 and "s" or "") .. " that you started" ..
+				" " .. (#started > 1 and "are" or "is") .. " " .. text, cid)
 		else
 			npcHandler:say("You haven't started any task yet.", cid)
 		end
 	elseif msg:lower() == "cancel" then
 		local started = player:getStartedTasks()
-		local task = getTaskByName(msg)
 		local text = ""
 		local sep = ", "
 		table.sort(started, (function(a, b) return (a < b) end))
@@ -366,24 +366,35 @@ local function creatureSayCallback(cid, type, msg)
 			text = text .. "{" .. (tasks[id].name or tasks[id].raceName) .. "}" .. sep
 		end
 		if started and #started > 0 then
-			npcHandler:say("Canceling a task will make the counter restart. Which of these tasks you want cancel?" .. (#started > 1 and "" or "") .. " " .. text, cid)
+			npcHandler:say("Canceling a task will make the counter restart. " ..
+						"Which of these tasks you want cancel?" .. (#started > 1 and "" or "") .. " " .. text, cid)
 			npcHandler.topic[cid] = 2
 		else
 			npcHandler:say("You haven't started any task yet.", cid)
 		end
-	elseif getTaskByName(msg) and npcHandler.topic[cid] == 2 and isInArray(getPlayerStartedTasks(cid), getTaskByName(msg)) then
+	elseif ((getTaskByName(msg)) and
+			(npcHandler.topic[cid] == 2) and
+			(isInArray(getPlayerStartedTasks(cid), getTaskByName(msg)))) then
 		local task = getTaskByName(msg)
 		if player:getStorageValue(KILLSSTORAGE_BASE + task) > 0 then
-			npcHandler:say("You currently killed " .. player:getStorageValue(KILLSSTORAGE_BASE + task) .. "/" .. tasks[task].killsRequired .. " " .. tasks[task].raceName .. ". Canceling this task will restart the count. Are you sure you want to cancel this task?", cid)
+			npcHandler:say("You currently killed " .. player:getStorageValue(KILLSSTORAGE_BASE + task) .. "/" ..
+						tasks[task].killsRequired .. " " .. tasks[task].raceName .. "."..
+						" ".."Canceling this task will restart the count."..
+						" ".."Are you sure you want to cancel this task?", cid)
 		else
 			npcHandler:say("Are you sure you want to cancel this task?", cid)
 		end
 		npcHandler.topic[cid] = 3
 		cancel[cid] = task
-	elseif getTaskByName(msg) and npcHandler.topic[cid] == 1 and isInArray(getPlayerStartedTasks(cid), getTaskByName(msg)) then
+	elseif ((getTaskByName(msg)) and
+			(npcHandler.topic[cid] == 1) and
+			(isInArray(getPlayerStartedTasks(cid), getTaskByName(msg)))) then
 		local task = getTaskByName(msg)
 		if player:getStorageValue(KILLSSTORAGE_BASE + task) > 0 then
-			npcHandler:say("You currently killed " .. player:getStorageValue(KILLSSTORAGE_BASE + task) .. "/" .. tasks[task].killsRequired .. " " .. tasks[task].raceName .. ".", cid)
+			npcHandler:say("You currently killed " ..
+						player:getStorageValue(KILLSSTORAGE_BASE + task) .. "/" ..
+						tasks[task].killsRequired .. " " ..
+						tasks[task].raceName .. ".", cid)
 		else
 			npcHandler:say("You currently killed 0/" .. tasks[task].killsRequired .. " " .. tasks[task].raceName .. ".", cid)
 		end
@@ -391,38 +402,49 @@ local function creatureSayCallback(cid, type, msg)
 	elseif msg:lower() == "yes" and npcHandler.topic[cid] == 3 then
 		player:setStorageValue(QUESTSTORAGE_BASE + cancel[cid], -1)
 		player:setStorageValue(KILLSSTORAGE_BASE + cancel[cid], -1)
-		npcHandler:say("You have canceled the task " .. (tasks[cancel[cid]].name or tasks[cancel[cid]].raceName) .. ".", cid)
+		npcHandler:say("You have canceled the task " ..
+					(tasks[cancel[cid]].name or tasks[cancel[cid]].raceName) .. ".", cid)
 		npcHandler.topic[cid] = 0
 	elseif isInArray({"points", "rank"}, msg:lower()) then
-		if player:getPawAndFurPoints() < 1 then
-			npcHandler:say("At this time, you have " .. player:getPawAndFurPoints() .. " Paw & Fur points. You " .. (player:getPawAndFurRank() == 6 and "are an Elite Hunter" or player:getPawAndFurRank() == 5 and "are a Trophy Hunter" or player:getPawAndFurRank() == 4 and "are a Big Game Hunter" or player:getPawAndFurRank() == 3 and "are a Ranger" or player:getPawAndFurRank() == 2 and "are a Huntsman" or player:getPawAndFurRank() == 1 and "are a Member"  or "haven't been ranked yet") .. ".", cid)
-		else
-			npcHandler:say("At this time, you have " .. player:getPawAndFurPoints() .. " Paw & Fur points. You " .. (player:getPawAndFurRank() == 6 and "are an Elite Hunter" or player:getPawAndFurRank() == 5 and "are a Trophy Hunter" or player:getPawAndFurRank() == 4 and "are a Big Game Hunter" or player:getPawAndFurRank() == 3 and "are a Ranger" or player:getPawAndFurRank() == 2 and "are a Huntsman" or player:getPawAndFurRank() == 1 and "are a Member"  or "haven't been ranked yet") .. ".", cid)
-		end
+		npcHandler:say("At this time, you have " .. player:getPawAndFurPoints() .. " Paw & Fur points. You " ..
+					(player:getPawAndFurRank() == 6 and "are an Elite Hunter" or
+					player:getPawAndFurRank() == 5 and "are a Trophy Hunter" or
+					player:getPawAndFurRank() == 4 and "are a Big Game Hunter" or
+					player:getPawAndFurRank() == 3 and "are a Ranger" or
+					player:getPawAndFurRank() == 2 and "are a Huntsman" or
+					player:getPawAndFurRank() == 1 and "are a Member"  or
+					"haven't been ranked yet") .. ".", cid)
 		npcHandler.topic[cid] = 0
 	elseif isInArray({"special task"}, msg:lower()) then
-		if player:getPawAndFurPoints() >= 70 then -- Tiquandas Revenge 70 points
-			if player:getStorageValue(Storage.KillingInTheNameOf.MissionTiquandasRevenge) == 1 then  -- Check if he has already started the task.
-				if (player:getStorageValue(Storage.KillingInTheNameOf.TiquandasRevengeTeleport) ~= 0) then
-					npcHandler:say("You have already started the task. Go find Tiquandas Revenge and take revenge yourself!", cid)
-				end
+		if (player:getPawAndFurPoints() >= 70) then
+			if ((player:getLevel() > 90) and
+				(player:getStorageValue(Storage.KillingInTheNameOf.MissionTiquandasRevenge) == 1) and
+				(player:getStorageValue(Storage.KillingInTheNameOf.TiquandasRevengeTeleport) ~= 0)) then
+				npcHandler:say("You have already started the task. Go find Tiquandas Revenge and take revenge yourself!", cid)
 			else
-				npcHandler:say({
-					"Have you heard about Tiquandas Revenge? It is said that the jungle itself is alive and takes revenge for all the bad things people have done to it. ...",
-					"I myself believe that there is some truth in this clap trap. Something 'real' which therefore must have a hideout somewhere. Go find it and take revenge yourself!"
-				}, cid)
-				player:setStorageValue(Storage.KillingInTheNameOf.TiquandasRevengeTeleport, 1) -- Task needed to enter Tiquandas Revenge TP
-				player:setStorageValue(Storage.KillingInTheNameOf.MissionTiquandasRevenge, 1) -- Won't give this task again.
+				npcHandler:say({"Have you heard about Tiquandas Revenge? " ..
+							"It is said that the jungle itself is alive and takes revenge for all the bad things people have done to it." ..
+							"...",
+							"I myself believe that there is some truth in this clap trap." ..
+							"Something 'real' which therefore must have a hideout somewhere." ..
+							"Go find it and take revenge yourself!"}, cid)
+				player:setStorageValue(Storage.KillingInTheNameOf.TiquandasRevengeTeleport, 1)
+				player:setStorageValue(Storage.KillingInTheNameOf.MissionTiquandasRevenge, 1)
 			end
 
-			if player:getStorageValue(Storage.KillingInTheNameOf.MissionDemodras) == 1 then  -- Check if he has already started the task.
-				if (player:getStorageValue(Storage.KillingInTheNameOf.DemodrasTeleport) ~= 0) then
-					npcHandler:say("You have already started the special task. Find Demodras and kill it.", cid)
-				end
+			if ((player:getLevel() > 100) and
+				(player:getStorageValue(Storage.KillingInTheNameOf.MissionDemodras) == 1) and
+				(player:getStorageValue(Storage.KillingInTheNameOf.DemodrasTeleport) ~= 0)) then
+				npcHandler:say("You have already started the special task. Find Demodras and kill it.", cid)
 			else
-				npcHandler:say("This task is a very dangerous one. I want you to look for Demodras' hideout. It might be somewhere under the Plains of Havoc. Good luck, old chap.", cid)
-				player:setStorageValue(Storage.KillingInTheNameOf.DemodrasTeleport, 1) -- Task needed to enter Demodras TP
-				player:setStorageValue(Storage.KillingInTheNameOf.MissionDemodras, 1) -- Won't give this task again.
+				if ((player:getStorageValue(Storage.KillingInTheNameOf.MissionTiquandasRevenge) == 1) and
+					(player:getStorageValue(Storage.KillingInTheNameOf.TiquandasRevengeTeleport) == 0)) then
+					npcHandler:say("This task is a very dangerous one. I want you to look for Demodras' hideout. "..
+								"It might be somewhere under the Plains of Havoc. Good luck, old chap.", cid)
+					player:setStorageValue(Storage.KillingInTheNameOf.DemodrasTeleport, 1)
+					player:setStorageValue(Storage.KillingInTheNameOf.MissionDemodras, 1)
+
+				end
 			end
 
 			if (player:getStorageValue(Storage.KillingInTheNameOf.MissionDemodras) == 1 and
@@ -431,8 +453,8 @@ local function creatureSayCallback(cid, type, msg)
 				player:getStorageValue(Storage.KillingInTheNameOf.TiquandasRevengeTeleport) == 0) then
 				npcHandler:say("You have already finished all special tasks.", cid)
 			end
+			npcHandler.topic[cid] = 0
 		end
-		npcHandler.topic[cid] = 0
 	end
 end
